fix: 타입 에러 해결 (#264)
